Refactored commom functions into different files
This was done for maintainability and reusabiity. Functions were generalized so that could be used for future projects.

arc_pyspark.py - contains functions that are related to Spark on ARC at the University of Calgary
file_py - contains functions that are related to IO
query.py - contains functions that are related to API query

Signed-off-by: Albert Leung <1319401+alby1976@users.noreply.github.com>