Able to create Article object
Added stubs to Crossref.py so that the corresponding
database can be queried and the objects related to
Publications can be instantiated and the relationships
between objects can be visualized. In addition, the
comparisons between the initial and final version of each
publication and be made.

Article is one the building blocks since it contains
information about each publication and prepublication.

Signed-off-by: Albert Leung <1319401+alby1976@users.noreply.github.com>