Implement PubPrepubArticleMediator
Using test-driven development this functionality was implement. Since
not all the  article in preprint service have been published so a special
list that contains to the doi to the published article. ArticleFactory stores
the prepublication doi and there is no easy way to receive the corrsponding
article from the doi of the published article. With this class one will be able
to retrieve the corresponding article.

NB: PublishedPrepubArticleMediator stores the earliest version of the
prepublished article. Currently, this may not be the initial version of the
article but will be rectified in the future.

Signed-off-by: Albert Leung <1319401+alby1976@users.noreply.github.com>