Implement pdf_manipulation.py
Using test-driven development this functionality was implement. With this
implementation one is able to extract pdf elements and text which is
need to do the comparison between the published article and its initial
version.

An issue occurred with conversion from pdf to txt. There are times when
the converted txt does not match the paragraph flow found in the pdf
document. Looking at different solutions to resolve this issue.

Signed-off-by: Albert Leung <1319401+alby1976@users.noreply.github.com>