Implement ConfigObj module in file.py
Using test-driven development this functionality was implement. With this
implementation one is able to extract create config file that is can be
read by ConfigObj module. With this ability one would be able to remove
different setting from code and into a file. This functionality improves
code maintainability and private setting from other people.

Signed-off-by: Albert Leung <1319401+alby1976@users.noreply.github.com>